requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-auth==2.6.2
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.34.2
google-cloud-bigtable==2.7.1
google-cloud-core==2.2.3
google-cloud-logging==3.0.0
google-cloud-pubsub==2.11.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
packaging==21.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
